添加商品时显示库存数

+++ app/src/main/java/com/zaomeng/zaomeng/model/repository/http/bean/GoodsDetailsBean.java
@@ -35,8 +35,8 @@ public class GoodsDetailsBean {
     private String littleImage;
     private double standPrice;
     private double realPrice;
-    private double stockNumber;
-    private double stockOut;
+    private int stockNumber;
+    private int stockOut;
     private String salesUnit;
     private String brandName;
     private double showPrice;
@@ -123,19 +123,19 @@ public class GoodsDetailsBean {
         this.realPrice = realPrice;
     }
 
-    public double getStockNumber() {
+    public int getStockNumber() {
         return stockNumber;
     }
 
-    public void setStockNumber(double stockNumber) {
+    public void setStockNumber(int stockNumber) {
         this.stockNumber = stockNumber;
     }
 
-    public double getStockOut() {
+    public int getStockOut() {
         return stockOut;
     }
 
-    public void setStockOut(double stockOut) {
+    public void setStockOut(int stockOut) {
         this.stockOut = stockOut;
     }
